Removes Q from the list of chars printed.
There seems to be some odd interaction with the Q letter.
It has been decided to remove Q from the markerset.
It acts weird and it seems to stem from outside the plugin.
It look very similar to O and can be hard to distinguish.
Q, you will be sorely missed.