Actions now mark their intent
TODO:
Complete rearchitechture:
instead of setupSecondOverlay and onActionPerformed there should be a
stack of "escapables". Whenever an overlay has been "finished" it will
push an escapable to the stack that can recreate the state right before
that overlay was finished. You can then have as many overlays as you want
and still be able to escape them off one at a time.

Additionally it would be neat with the ability to "up" an intent. fx if
you wish to jump to some location and you fatfinger the jump and the
intent states the one next where you want to go then it would be neat to
be able to press up and then go to next marker instead.